refactor: extract navigation from App
