+++ src/app/use-app-navigation.ts
@@ -1,0 +1,262 @@
+import { useCallback, useEffect, useMemo, useRef } from "react";
+import {
+  AGENT_NEW_SESSION_EVENT,
+  markAgentNewSessionPending,
+  type AgentNewSessionDetail,
+} from "../components/agent/AgentWorkspace";
+import { defaultNav, makeTabId, navEquals, type TabNav } from "./tabs/tabs";
+import { getNote } from "../lib/tauri";
+import { messageFromError } from "../lib/errors";
+import { agentSessionTabTitle, refreshedTabNav } from "./app-shell";
+import type { UseAppNavigationDependencies } from "./use-app-navigation-types";
+
+export function useAppNavigation(dependencies: UseAppNavigationDependencies) {
+  const {
+    activeAgentSessionId,
+    activeAgentSessionSeed,
+    activeTabId,
+    activeTabIdRef,
+    activeView,
+    activeViewRef,
+    agentOrigin,
+    agentSessions,
+    dispatch,
+    originAllNotes,
+    originFolderId,
+    pendingSessionProjectRef,
+    restoreTargetRef,
+    selectedNoteId,
+    setActiveAgentSession,
+    setActiveAgentSessionId,
+    setActiveAgentSessionSeed,
+    setActiveTabId,
+    setActiveView,
+    setAgentOrigin,
+    setError,
+    setFolderReturnTarget,
+    setOriginAllNotes,
+    setOriginFolderId,
+    setSettingsReturnView,
+    setTabs,
+    state,
+    tabs,
+    tabsRef,
+  } = dependencies;
+
+  const liveNav = useMemo<TabNav>(
+    () => ({
+      view: activeView,
+      noteId: activeView === "meetings" ? selectedNoteId : undefined,
+      originFolderId: activeView === "meetings" ? originFolderId : undefined,
+      originAllNotes: activeView === "meetings" ? originAllNotes : undefined,
+      folderId: activeView === "folders" ? state.selectedFolderId : undefined,
+      agentSessionId: activeView === "agent" ? activeAgentSessionId : undefined,
+      agentSessionTitle:
+        activeView === "agent" ? agentSessionTabTitle(activeAgentSessionSeed) : undefined,
+      agentOrigin: activeView === "agent" ? agentOrigin : undefined,
+    }),
+    [
+      activeView,
+      selectedNoteId,
+      originFolderId,
+      originAllNotes,
+      state.selectedFolderId,
+      activeAgentSessionId,
+      activeAgentSessionSeed?.preview,
+      activeAgentSessionSeed?.title,
+      agentOrigin,
+    ],
+  );
+
+  // Mirror live navigation into the active tab. While a restore is in flight we
+  // hold off until live nav settles onto the target, then release — this keeps
+  // an async note load mid-switch from stamping a half-built snapshot onto the
+  // tab we're moving to.
+  useEffect(() => {
+    if (restoreTargetRef.current) {
+      if (navEquals(restoreTargetRef.current, liveNav)) {
+        restoreTargetRef.current = null;
+      }
+      return;
+    }
+    setTabs((prev) =>
+      prev.map((tab) => {
+        if (tab.id !== activeTabId) return tab;
+        const nav = refreshedTabNav(tab.nav, liveNav);
+        return nav ? { ...tab, nav } : tab;
+      }),
+    );
+  }, [liveNav, activeTabId]);
+
+  useEffect(() => {
+    tabsRef.current = tabs;
+  }, [tabs]);
+
+  useEffect(() => {
+    activeTabIdRef.current = activeTabId;
+  }, [activeTabId]);
+
+  // Keep the latest selected note id reachable from applyNav without making it
+  // a dependency (which would rebuild the callback on every note change).
+  const selectedNoteIdRef = useRef(selectedNoteId);
+  useEffect(() => {
+    selectedNoteIdRef.current = selectedNoteId;
+  }, [selectedNoteId]);
+
+  // Push a snapshot into live state. Used by tab switch / open only; in-tab
+  // navigation keeps flowing through the existing handlers untouched.
+  const applyNav = useCallback(
+    (nav: TabNav) => {
+      // The guard only needs to bridge the async note fetch — everything else
+      // applies synchronously. So hold capture only while a fetch is in flight
+      // and release it when the fetch settles (success or failure), never tying
+      // release to a target that might be unreachable (a deleted session/note).
+      restoreTargetRef.current = nav;
+      setAgentOrigin(nav.view === "agent" ? nav.agentOrigin : undefined);
+      setOriginFolderId(nav.view === "meetings" ? nav.originFolderId : undefined);
+      setOriginAllNotes(nav.view === "meetings" ? !!nav.originAllNotes : false);
+      // The "back to <note>" breadcrumb target isn't part of a tab's snapshot,
+      // so clear it on every restore — otherwise it leaks from the tab that set
+      // it into whatever tab we switch to.
+      setFolderReturnTarget(undefined);
+      if (nav.view === "folders") {
+        dispatch({ type: "folderSelected", folderId: nav.folderId });
+      }
+      // Mirror openSettings: a settings tab (e.g. cmd-clicked open) needs a
+      // return view recorded so exiting Settings lands where it came from.
+      if (nav.view === "settings") {
+        const returnView = activeViewRef.current;
+        if (returnView !== "settings") setSettingsReturnView(returnView);
+      }
+      if (nav.view === "agent") {
+        const session = nav.agentSessionId
+          ? (agentSessions.find((s) => s.id === nav.agentSessionId) ?? {
+              id: nav.agentSessionId,
+              title: nav.agentSessionTitle,
+            })
+          : undefined;
+        setActiveAgentSessionId(nav.agentSessionId);
+        setActiveAgentSessionSeed(session);
+      } else {
+        setActiveAgentSession(undefined);
+      }
+      const needsNoteLoad =
+        nav.view === "meetings" && !!nav.noteId && selectedNoteIdRef.current !== nav.noteId;
+      if (needsNoteLoad) {
+        const noteId = nav.noteId!;
+        void getNote(noteId)
+          .then((note) => dispatch({ type: "noteLoaded", note }))
+          .catch((err: unknown) => setError(messageFromError(err)))
+          .finally(() => {
+            // A newer restore may have superseded this one — only release the
+            // guard if it's still ours.
+            if (restoreTargetRef.current === nav) {
+              restoreTargetRef.current = null;
+            }
+          });
+      } else {
+        // Nothing async to wait for — let capture resume immediately.
+        restoreTargetRef.current = null;
+      }
+      setActiveView(nav.view);
+    },
+    [agentSessions],
+  );
+  const applyNavRef = useRef(applyNav);
+  useEffect(() => {
+    applyNavRef.current = applyNav;
+  }, [applyNav]);
+
+  function activateTab(id: string) {
+    if (id === activeTabId) return;
+    const target = tabs.find((tab) => tab.id === id);
+    if (!target) return;
+    setActiveTabId(id);
+    applyNav(target.nav);
+  }
+
+  // Open a fresh tab on the given snapshot and focus it. The active tab's own
+  // snapshot was already captured by the mirror effect, so nothing is lost.
+  function openTab(nav: TabNav) {
+    const id = makeTabId();
+    const index = tabs.findIndex((tab) => tab.id === activeTabId);
+    const next = [...tabs];
+    // Drop the new tab in just to the right of the active one, like a browser.
+    next.splice(index < 0 ? tabs.length : index + 1, 0, { id, nav });
+    setTabs(next);
+    setActiveTabId(id);
+    applyNav(nav);
+  }
+
+  // Drive live state to a brand-new chat: arm the new-session handshake so the
+  // agent workspace opens on the hero instead of restoring the last
+  // conversation. Mirrors handleNewAgentSession (applyNav alone would only swap
+  // state, leaving the previous chat on screen under a "New chat" label).
+  const armNewChatLive = useCallback(() => {
+    restoreTargetRef.current = { view: "agent" };
+    pendingSessionProjectRef.current = null;
+    setAgentOrigin(undefined);
+    markAgentNewSessionPending();
+    setActiveAgentSession(undefined);
+    setActiveView("agent");
+    window.setTimeout(() => {
+      window.dispatchEvent(new CustomEvent<AgentNewSessionDetail>(AGENT_NEW_SESSION_EVENT));
+    }, 0);
+  }, []);
+
+  // The "+" / ⌘T affordance: a new tab is always a fresh chat.
+  function openNewChatTab() {
+    const id = makeTabId();
+    const index = tabs.findIndex((tab) => tab.id === activeTabId);
+    const next = [...tabs];
+    next.splice(index < 0 ? tabs.length : index + 1, 0, {
+      id,
+      nav: defaultNav(),
+    });
+    setTabs(next);
+    setActiveTabId(id);
+    armNewChatLive();
+  }
+
+  const closeTab = useCallback(
+    (id: string) => {
+      const currentTabs = tabsRef.current;
+      const currentActiveTabId = activeTabIdRef.current;
+
+      if (currentTabs.length <= 1) {
+        // Never leave the strip empty — reset the sole tab to a fresh chat.
+        const fresh = { id: makeTabId(), nav: defaultNav() };
+        tabsRef.current = [fresh];
+        activeTabIdRef.current = fresh.id;
+        setTabs([fresh]);
+        setActiveTabId(fresh.id);
+        armNewChatLive();
+        return;
+      }
+      const index = currentTabs.findIndex((tab) => tab.id === id);
+      if (index < 0) return;
+      const next = currentTabs.filter((tab) => tab.id !== id);
+      tabsRef.current = next;
+      setTabs(next);
+      if (id === currentActiveTabId) {
+        // Focus the right neighbor, falling back to the left — browser behavior.
+        const neighbor = next[index] ?? next[index - 1];
+        if (neighbor) {
+          activeTabIdRef.current = neighbor.id;
+          setActiveTabId(neighbor.id);
+          applyNavRef.current(neighbor.nav);
+        }
+      }
+    },
+    [armNewChatLive],
+  );
+
+  return {
+    applyNav,
+    activateTab,
+    openTab,
+    openNewChatTab,
+    closeTab,
+    selectedNoteIdRef,
+  };
+}
